+++ root/programs/CS/Frameworks/Infrastructure/Framework/Presentation/BaseController.cs
@@ -109,6 +109,7 @@ using System.Web.UI.WebControls;
 using Touryo.Infrastructure.Framework.Exceptions;
 using Touryo.Infrastructure.Framework.Util;
 using Touryo.Infrastructure.Public.IO;
+using Touryo.Infrastructure.Public.Str;
 using Touryo.Infrastructure.Public.Util;
 
 namespace Touryo.Infrastructure.Framework.Presentation
