+++ root/programs/CS/Frameworks/Infrastructure/Framework/Transmission/InProcessNameService.cs
@@ -46,6 +46,7 @@ using System.Xml;
 using Touryo.Infrastructure.Framework.Exceptions;
 using Touryo.Infrastructure.Framework.Util;
 using Touryo.Infrastructure.Public.IO;
+using Touryo.Infrastructure.Public.Str;
 using Touryo.Infrastructure.Public.Util;
 
 namespace Touryo.Infrastructure.Framework.Transmission
