+++ root/programs/CS/Frameworks/Infrastructure/Framework/Transmission/ProtocolNameService.cs
@@ -48,6 +48,7 @@ using System.Collections.Generic;
 using Touryo.Infrastructure.Framework.Exceptions;
 using Touryo.Infrastructure.Framework.Util;
 using Touryo.Infrastructure.Public.IO;
+using Touryo.Infrastructure.Public.Str;
 using Touryo.Infrastructure.Public.Util;
 
 namespace Touryo.Infrastructure.Framework.Transmission
