+++ root/programs/CS/Frameworks/Infrastructure/Framework/Util/GetMessage.cs
@@ -52,6 +52,7 @@ using System.Collections.Generic;
 
 using Touryo.Infrastructure.Framework.Exceptions;
 using Touryo.Infrastructure.Public.IO;
+using Touryo.Infrastructure.Public.Str;
 using Touryo.Infrastructure.Public.Util;
 
 namespace Touryo.Infrastructure.Framework.Util
