+++ root/programs/CS/Frameworks/Infrastructure/Framework/Util/GetSharedProperty.cs
@@ -40,6 +40,7 @@ using System.Collections.Generic;
 
 using Touryo.Infrastructure.Framework.Exceptions;
 using Touryo.Infrastructure.Public.IO;
+using Touryo.Infrastructure.Public.Str;
 using Touryo.Infrastructure.Public.Util;
 
 namespace Touryo.Infrastructure.Framework.Util
