+++ root/programs/CS/Frameworks/Infrastructure/Public/Public_net47.csproj
@@ -138,7 +138,7 @@
     <Compile Include="Db\DbEnum.cs" />
     <Compile Include="Str\ToStringExtension.cs" />
     <Compile Include="Util\StackFrameOperator.cs" />
-    <Compile Include="Util\StringVariableOperator.cs" />
+    <Compile Include="Str\StringVariableOperator.cs" />
     <Compile Include="Util\CustomMarshaler.cs" />
     <Compile Include="FastReflection\CompiledExpressionCreater.cs" />
     <Compile Include="Util\ArrayOperator.cs" />
