+++ root/programs/CS/Samples/Bat_sample/SimpleBatch_sample/Program.cs
@@ -25,6 +25,7 @@ using System.Collections.Generic;
 
 using Touryo.Infrastructure.Business.Util;
 using Touryo.Infrastructure.Public.Db;
+using Touryo.Infrastructure.Public.Str;
 using Touryo.Infrastructure.Public.Util;
 
 namespace SimpleBatch_sample
