--- root/programs/CS/y_Build_TestCode_SecCUI.txt
@@ -1,13 +1,0 @@
-以下をWSL(Windows Subsystem for Linux)から実行.
-
-<EncAndDecUtilCUI(20)>
-cd /mnt/c/Git1/OpenTouryo/root/programs/CS/Frameworks/Tests/EncAndDecUtilCUI/core20
-dotnet publish -c Release -r ubuntu.16.04-x64 --self-contained
-cd bin/Release/netcoreapp2.0/ubuntu.16.04-x64
-dotnet EncAndDecUtilCUICore.dll > ../../../../../ResultCore20OnLinux.txt
-
-<EncAndDecUtilCUI(30)>
-cd /mnt/c/Git1/OpenTouryo/root/programs/CS/Frameworks/Tests/EncAndDecUtilCUI/core30
-dotnet publish -c Release -r ubuntu.16.04-x64 --self-contained
-cd bin/Release/netcoreapp3.0/ubuntu.16.04-x64
-dotnet EncAndDecUtilCUICore.dll > ../../../../../ResultCore30OnLinux.txt
